+++ src/ts-main/errors/WaspOtherError.ts
@@ -15,8 +15,6 @@ export default class WaspOtherError extends Error {
   /** Specifies the exact error encountered. */
   public readonly code:
     | "MediaSourceAttachmentError"
-    | "NotAPlaylist"
-    | "NoSupportedVariant"
     | "UnfoundLockedVariant"
     | "Unknown";
 
@@ -46,12 +44,6 @@ export default class WaspOtherError extends Error {
         break;
       case OtherErrorCode.UnfoundLockedVariant:
         this.code = WaspErrorCode.UnfoundLockedVariant;
-        break;
-      case OtherErrorCode.NoSupportedVariant:
-        this.code = WaspErrorCode.NoSupportedVariant;
-        break;
-      case OtherErrorCode.NotAPlaylist:
-        this.code = WaspErrorCode.NotAPlaylist;
         break;
       default:
         this.code = WaspErrorCode.Unknown;